+++ vim/.config/nvim/init.lua
@@ -1,0 +1,150 @@
+-- Basic settings
+vim.g.mapleader = " "
+vim.g.maplocalleader = " "
+
+-- Enable line numbers and relative line numbers
+vim.wo.number = true
+vim.wo.relativenumber = true
+
+-- Enable mouse support
+vim.o.mouse = "a"
+
+-- Enable true colors
+vim.o.termguicolors = true
+
+-- Set tab settings
+vim.o.tabstop = 2
+vim.o.shiftwidth = 2
+vim.o.expandtab = true
+vim.o.smartindent = true
+
+-- Enable clipboard
+vim.o.clipboard = "unnamedplus"
+
+-- Enable incremental search
+vim.o.incsearch = true
+
+-- Bootstrap lazy.nvim
+local lazypath = vim.fn.stdpath("data") .. "/lazy/lazy.nvim"
+if not vim.loop.fs_stat(lazypath) then
+    vim.fn.system({
+        "git",
+        "clone",
+        "--filter=blob:none",
+        "https://github.com/folke/lazy.nvim.git",
+        "--branch=stable",
+        lazypath,
+    })
+end
+vim.opt.rtp:prepend(lazypath)
+
+-- Initialize lazy.nvim
+require("lazy").setup({
+    -- Theme
+    require("theme"),
+
+    -- Completion
+    {
+        "hrsh7th/nvim-cmp",
+        dependencies = {
+            "hrsh7th/cmp-buffer",
+            "hrsh7th/cmp-path",
+            "L3MON4D3/LuaSnip",
+            "saadparwaiz1/cmp_luasnip",
+            "rafamadriz/friendly-snippets",
+        },
+    },
+
+    -- Telescope for fuzzy finding
+    {
+        "nvim-telescope/telescope.nvim",
+        dependencies = {
+            "nvim-lua/plenary.nvim",
+        },
+    },
+
+    -- Treesitter for better syntax highlighting
+    {
+        "nvim-treesitter/nvim-treesitter",
+        build = ":TSUpdate",
+    },
+
+    -- Git integration
+    "lewis6991/gitsigns.nvim",
+
+    -- Status line
+    "nvim-lualine/lualine.nvim",
+
+    -- File explorer
+    {
+        "nvim-tree/nvim-tree.lua",
+        dependencies = {
+            "nvim-tree/nvim-web-devicons",
+        },
+        config = function()
+            require("nvim-tree").setup({
+                view = {
+                    width = 30,
+                },
+                renderer = {
+                    group_empty = true,
+                },
+                filters = {
+                    dotfiles = false,
+                },
+            })
+        end,
+    },
+
+    -- Buffer line for better tab visualization
+    {
+        "akinsho/bufferline.nvim",
+        version = "*",
+        dependencies = "nvim-tree/nvim-web-devicons",
+        config = function()
+            require("bufferline").setup({
+                options = {
+                    mode = "buffers",
+                    separator_style = "slant",
+                    always_show_bufferline = true,
+                    show_buffer_close_icons = true,
+                    show_close_icon = true,
+                    color_icons = true,
+                    offsets = {
+                        {
+                            filetype = "NvimTree",
+                            text = "File Explorer",
+                            text_align = "center",
+                            separator = true,
+                        },
+                    },
+                },
+                highlights = require("catppuccin.groups.integrations.bufferline").get(),
+            })
+        end,
+    },
+})
+
+-- Load completion configuration
+require("completion")
+
+-- Set colorscheme after lazy.nvim loads
+vim.cmd.colorscheme("catppuccin")
+
+-- Key mappings
+-- File explorer
+vim.keymap.set("n", "<leader>e", ":NvimTreeToggle<CR>")
+vim.keymap.set("n", "<leader>o", ":NvimTreeFocus<CR>")
+
+-- Telescope
+vim.keymap.set("n", "<leader>ff", ":Telescope find_files<CR>")
+vim.keymap.set("n", "<leader>fg", ":Telescope live_grep<CR>")
+vim.keymap.set("n", "<leader>fb", ":Telescope buffers<CR>")
+
+-- Buffer navigation
+vim.keymap.set("n", "<Tab>", ":BufferLineCycleNext<CR>")
+vim.keymap.set("n", "<S-Tab>", ":BufferLineCyclePrev<CR>")
+vim.keymap.set("n", "<leader>bc", ":BufferLinePickClose<CR>")
+vim.keymap.set("n", "<leader>bp", ":BufferLinePick<CR>")
+vim.keymap.set("n", "<leader>bl", ":BufferLineCloseLeft<CR>")
+vim.keymap.set("n", "<leader>br", ":BufferLineCloseRight<CR>")